WIP:Move LayoutSlip drawing code into LayoutSlipView
